Refactor: move ZigException functions to dedicated file
Move error-related functions from bindings.cpp into a new ZigException.cpp file for better code organization:

- populateStackFrameMetadata
- populateStackFramePosition
- populateStackFrame
- populateStackTrace
- fromErrorInstance
- exceptionFromString
- JSC__JSValue__toZigException
- ZigException__collectSourceLines
- JSC__Exception__getStackTrace

Also moved helper functions and types:
- V8StackTraceIterator class
- getNonObservable
- PopulateStackTraceFlags enum
- StringView_slice helper
- SYNTAX_ERROR_CODE macro

This improves code organization by grouping all exception handling logic together in a dedicated file.
